refactor: Rename Calendar component to MenuCalendar and update its implementation for better clarity and functionality
